+++ packages/material/modules/widgets/value/MaterialAutocomplete.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, {useCallback} from "react";
 import omit from "lodash/omit";
 import TextField from "@material-ui/core/TextField";
 import FormControl from "@material-ui/core/FormControl";
@@ -14,7 +14,6 @@ const { useListValuesAutocomplete } = Hooks;
 
 const nonCheckedIcon = <CheckBoxOutlineBlankIcon fontSize="small" style={{ marginRight: 10, marginTop: 4 }} />;
 const checkedIcon = <CheckBoxIcon fontSize="small" style={{ marginRight: 10, marginTop: 4 }} />;
-const defaultFilterOptions = createFilterOptions();
 const emptyArray = [];
 
 
@@ -23,7 +22,17 @@ export default (props) => {
     allowCustomValues, multiple, disableClearable,
     value: selectedValue, customProps, readonly, config, filterOptionsConfig, errorText
   } = props;
-  const filterOptionsFn = filterOptionsConfig ? createFilterOptions(filterOptionsConfig) : defaultFilterOptions;
+  const stringifyOption = useCallback((option) => {
+    const keysForFilter = config.settings.listKeysForSearch;
+    const valueForFilter = keysForFilter
+      .map(k => (typeof option[k] == "string" ? option[k] : ""))
+      .join("\0");
+    return valueForFilter;
+  }, [config]);
+  const defaultFilterOptionsConfig = {
+    stringify: stringifyOption
+  };
+  const filterOptionsFn = createFilterOptions(filterOptionsConfig || defaultFilterOptionsConfig);
 
   // hook
   const {