+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Updated `@fluentui/react`. Fixed date/time widgets
   - Autocomplete items now support `groupTitle` (PR #953) (issue #600)
   - Add more types in `Utils.Autocomplete` (PR #953) (issue #934)
+  - Add settings `fieldItemKeysForSearch` and `listKeysForSearch` (PR #xxx) (issue #931)
 - 6.3.0
   - Allow saving and loading config from server (PR #866) (issue #817)
     - New utils: `compressConfig()`, `decompressConfig()`